+++ Week_11_TTMC_GPU/v3_gpu_5loop.cu
@@ -20,18 +20,18 @@ do {                                                                            
     }                                                                                \
 } while (0)
 
-// // atomicAdd_double is now defined here
-// #ifndef ATOMIC_ADD_DOUBLE_DEFINED
-// #define ATOMIC_ADD_DOUBLE_DEFINED
-// __device__ double atomicAdd_double(double* address, double val) {
+// // atomicAdd_float is now defined here
+// #ifndef ATOMIC_ADD_float_DEFINED
+// #define ATOMIC_ADD_float_DEFINED
+// __device__ float atomicAdd_float(float* address, float val) {
 //   unsigned long long int* address_as_ull = (unsigned long long int*)address;
 //   unsigned long long int old = *address_as_ull, assumed;
 //   do {
 //     assumed = old;
 //     old = atomicCAS(address_as_ull, assumed,
-//     __double_as_longlong(val + __longlong_as_double(assumed)));
+//     __float_as_longlong(val + __longlong_as_float(assumed)));
 //   } while (assumed != old);
-//   return __longlong_as_double(old);
+//   return __longlong_as_float(old);
 // }
 // #endif
 
@@ -41,8 +41,8 @@ __global__ void GPU_5loop_contraction_kernel_0(
   uint64_t* mode_0_ptr, uint64_t* mode_0_idx,
   uint64_t* mode_1_ptr, uint64_t* mode_1_idx,
   uint64_t* mode_2_ptr, uint64_t* mode_2_idx,
-  double* values, double* arr_A, double* arr_B,  
-  double* arr_O, uint64_t l, uint64_t m, uint64_t n, uint64_t f1, uint64_t f2, int contraction,
+  float* values, float* arr_A, float* arr_B,  
+  float* arr_O, uint64_t l, uint64_t m, uint64_t n, uint64_t f1, uint64_t f2, int contraction,
   uint64_t size_mode_0_ptr, uint64_t size_mode_1_ptr, uint64_t size_mode_2_ptr,
   uint64_t size_mode_0_idx, uint64_t size_mode_1_idx, uint64_t size_mode_2_idx) 
 {
@@ -58,7 +58,7 @@ __global__ void GPU_5loop_contraction_kernel_0(
     }
   }
   uint64_t i, j, k, index_A, index_B, index_O;
-  double value;
+  float value;
   if ((i_ptr >= 0 && i_ptr < mode_0_ptr[1]) && 
       ( j_ptr < size_mode_1_idx) ) 
   {
@@ -109,8 +109,8 @@ void performContraction_gpu_1(
   uint64_t* mode_0_ptr, uint64_t* mode_0_idx,
   uint64_t* mode_1_ptr, uint64_t* mode_1_idx,
   uint64_t* mode_2_ptr, uint64_t* mode_2_idx,
-  double* values, double* arr_A, double* arr_B,
-  double* arr_O, uint64_t arr_A_size, uint64_t arr_B_size, uint64_t arr_O_size,
+  float* values, float* arr_A, float* arr_B,
+  float* arr_O, uint64_t arr_A_size, uint64_t arr_B_size, uint64_t arr_O_size,
   int contraction, uint64_t dim_0, uint64_t dim_1, uint64_t dim_2,
   uint64_t r1, uint64_t r2, uint64_t total_values,
   uint64_t size_mode_0_ptr, uint64_t size_mode_1_ptr, uint64_t size_mode_2_ptr,
@@ -118,7 +118,7 @@ void performContraction_gpu_1(
 {
   // Allocate device memory
   uint64_t *d_mode_0_ptr, *d_mode_0_idx, *d_mode_1_ptr, *d_mode_1_idx, *d_mode_2_ptr, *d_mode_2_idx;
-  double *d_values, *d_arr_A, *d_arr_B, *d_arr_O;
+  float *d_values, *d_arr_A, *d_arr_B, *d_arr_O;
 
   cudaMalloc(&d_mode_0_ptr, sizeof(uint64_t) * size_mode_0_ptr);
   cudaMalloc(&d_mode_0_idx, sizeof(uint64_t) * size_mode_0_idx);
@@ -126,10 +126,10 @@ void performContraction_gpu_1(
   cudaMalloc(&d_mode_1_idx, sizeof(uint64_t) * size_mode_1_idx);
   cudaMalloc(&d_mode_2_ptr, sizeof(uint64_t) * size_mode_2_ptr);
   cudaMalloc(&d_mode_2_idx, sizeof(uint64_t) * size_mode_2_idx);
-  cudaMalloc(&d_values, sizeof(double) * total_values);
-  cudaMalloc(&d_arr_A, sizeof(double) * arr_A_size);
-  cudaMalloc(&d_arr_B, sizeof(double) * arr_B_size);
-  cudaMalloc(&d_arr_O, sizeof(double) * arr_O_size);
+  cudaMalloc(&d_values, sizeof(float) * total_values);
+  cudaMalloc(&d_arr_A, sizeof(float) * arr_A_size);
+  cudaMalloc(&d_arr_B, sizeof(float) * arr_B_size);
+  cudaMalloc(&d_arr_O, sizeof(float) * arr_O_size);
 
   // Copy data from host to device
   cudaMemcpy(d_mode_0_ptr, mode_0_ptr, sizeof(uint64_t) * size_mode_0_ptr, cudaMemcpyHostToDevice);
@@ -138,10 +138,10 @@ void performContraction_gpu_1(
   cudaMemcpy(d_mode_1_idx, mode_1_idx, sizeof(uint64_t) * size_mode_1_idx, cudaMemcpyHostToDevice);
   cudaMemcpy(d_mode_2_ptr, mode_2_ptr, sizeof(uint64_t) * size_mode_2_ptr, cudaMemcpyHostToDevice);
   cudaMemcpy(d_mode_2_idx, mode_2_idx, sizeof(uint64_t) * size_mode_2_idx, cudaMemcpyHostToDevice);
-  cudaMemcpy(d_values, values, sizeof(double) * total_values, cudaMemcpyHostToDevice);
-  cudaMemcpy(d_arr_A, arr_A, sizeof(double) * arr_A_size, cudaMemcpyHostToDevice);
-  cudaMemcpy(d_arr_B, arr_B, sizeof(double) * arr_B_size, cudaMemcpyHostToDevice);
-  cudaMemset(d_arr_O, 0, sizeof(double) * arr_O_size);
+  cudaMemcpy(d_values, values, sizeof(float) * total_values, cudaMemcpyHostToDevice);
+  cudaMemcpy(d_arr_A, arr_A, sizeof(float) * arr_A_size, cudaMemcpyHostToDevice);
+  cudaMemcpy(d_arr_B, arr_B, sizeof(float) * arr_B_size, cudaMemcpyHostToDevice);
+  cudaMemset(d_arr_O, 0, sizeof(float) * arr_O_size);
 
   // Kernel launch parameters
   int threadsPerBlock = 256;
@@ -160,7 +160,7 @@ void performContraction_gpu_1(
   cout << "Method: GPU_5L, Time: " << duration / 1000.0 << " ms" << endl;
 
   // Copy results back to host
-  cudaMemcpy(arr_O, d_arr_O, sizeof(double) * arr_O_size, cudaMemcpyDeviceToHost);
+  cudaMemcpy(arr_O, d_arr_O, sizeof(float) * arr_O_size, cudaMemcpyDeviceToHost);
 
   // Free device memory
   cudaFree(d_mode_0_ptr);
@@ -232,7 +232,7 @@ int main(int argc, char* argv[]) {
         uint64_t *mode_0_ptr, *mode_0_idx;
         uint64_t *mode_1_ptr, *mode_1_idx;
         uint64_t *mode_2_ptr, *mode_2_idx;
-        double *values;
+        float *values;
         int order;
         
         size_t size_mode_0_ptr = tensor.ptrs[0].size();
@@ -259,7 +259,7 @@ int main(int argc, char* argv[]) {
         uint64_t out_dim1 = getOutputDim1(dimensions, ncm);
         
         // Generate factor matrices
-        double *arr_A = nullptr, *arr_B = nullptr;
+        float *arr_A = nullptr, *arr_B = nullptr;
         generate_matrix(matrix_dim1, rank1, 42, arr_A);
         generate_matrix(matrix_dim2, rank2, 43, arr_B);
         
@@ -276,8 +276,8 @@ int main(int argc, char* argv[]) {
         }
         
         // Allocate output array
-        double* arr_O = allocate_aligned_array(arr_O_size);
-        double* ref_O = nullptr;
+        float* arr_O = allocate_aligned_array(arr_O_size);
+        float* ref_O = nullptr;
         
         if (verify) {
             // Only allocate reference array if verification is needed
@@ -306,7 +306,7 @@ int main(int argc, char* argv[]) {
         auto duration = std::chrono::duration_cast<std::chrono::microseconds>(end - start).count();
         
         bool valid = true;
-        double ref_duration = 0.0;
+        float ref_duration = 0.0;
         
         if (verify) {
             // Now run reference implementation (CPU 4-loop) for validation
